+++ app/build.gradle
@@ -8,14 +8,14 @@ android {
         applicationId "com.eli.orange"
         minSdkVersion 17
         targetSdkVersion 28
-        versionCode 1
+        versionCode 2
         versionName "0.1"
         multiDexEnabled true
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
@@ -82,7 +82,7 @@ dependencies {
     implementation 'com.firebaseui:firebase-ui-storage:3.3.0'
 
     //Add Google Material
-    implementation 'com.google.android.material:material:1.1.0-alpha06'
+    implementation 'com.google.android.material:material:1.1.0-alpha07'
 
 
     // MterialSearchViewShow
