+++ app/src/main/res/drawable/price_square.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android" >
 
-    <solid android:color="@color/primaryColor" />
+    <solid android:color="#FFFF00" />
 
     <stroke
         android:width="1dp"
-        android:color="@color/primaryColor" />
+        android:color="#FFFF00" />
     <corners
-        android:radius="10dp"
+        android:radius="4dp"
         />
 
 </shape>
