+++ app/src/main/res/layout/all_uploads_layout.xml
@@ -8,85 +8,78 @@
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:cardCornerRadius="4dp">
+        app:cardCornerRadius="1dp"
+        android:layout_marginTop="1dp"
+        app:cardBackgroundColor="#ffffff">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
-            android:layout_height="120dp">
-
-            <ImageButton
-                android:id="@+id/uploadsLocation"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:src="@drawable/crosshairs_gps"
-                android:translationZ="1dp"
-                android:padding="2dp"
-                android:visibility="invisible"
-                android:tooltipText="View Seller Location"
-                android:background="#00ffffff"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintTop_toTopOf="parent"/>
-
-
-            <TextView
-                android:id="@+id/textViewName"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Name"
-                android:visibility="gone"
-                android:textAppearance="?android:textAppearanceSmall"
-                app:layout_constraintTop_toBottomOf="@id/imageView"
-                app:layout_constraintEnd_toEndOf="@id/imageView"
-                app:layout_constraintStart_toStartOf="@id/imageView"/>
+            android:layout_height="wrap_content"
+            android:padding="4dp">
 
 
 
             <ImageView
                 android:id="@+id/imageView"
                 android:layout_width="@dimen/image_size"
-                android:layout_height="match_parent"
+                android:layout_height="@dimen/image_size"
                 android:scaleType="centerCrop"
-                android:layout_below="@id/textViewName"
                 android:src="@drawable/placeholder_image"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent"
                 app:layout_constraintBottom_toBottomOf="parent"/>
 
             <TextView
-                android:id="@+id/textViewprice"
+                android:id="@+id/textViewBuyNow"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/price"
+                android:text="@string/buy_now"
                 android:paddingVertical="2dp"
                 android:paddingHorizontal="@dimen/card_margin"
                 android:textAlignment="center"
-                android:textColor="@color/green"
+                android:textColor="@color/white"
+                android:textStyle="bold"
                 android:textAppearance="@style/TextAppearance.AppCompat.Small"
                 android:background="@drawable/filled_roundededges_box_dark"
                 app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent"
-                android:layout_margin="@dimen/card_margin"/>
+                app:layout_constraintTop_toBottomOf="@id/textViewDesc"
+                />
+            <TextView
+                android:id="@+id/textViewprice"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/buy_now"
+                android:paddingVertical="2dp"
+                android:paddingHorizontal="@dimen/card_margin"
+                android:textAlignment="center"
+                android:textColor="@color/red"
+                android:textStyle="bold"
+                android:textAppearance="@style/TextAppearance.AppCompat.Small"
+                android:background="@drawable/price_square"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toTopOf="@id/imageView"
+                />
 
 
             <TextView
                 android:id="@+id/textViewTitle"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/price"
+                android:text="@string/image_title"
                 android:padding="3dp"
+                android:textStyle="bold"
                 android:layout_marginStart="@dimen/card_margin"
                 android:textAppearance="@style/TextAppearance.AppCompat.Small"
                 app:layout_constraintStart_toEndOf="@id/imageView"
-                app:layout_constraintTop_toTopOf="parent"/>
+                app:layout_constraintTop_toTopOf="@id/imageView"/>
             <TextView
                 android:id="@+id/textViewDesc"
-                android:layout_width="match_parent"
+                android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/price"
-                android:padding="3dp"
-                android:layout_marginStart="@dimen/image_size"
+                android:text="@string/long_text_example"
+                android:maxLines="2"
                 android:textAppearance="@style/TextAppearance.AppCompat.Small"
-                app:layout_constraintStart_toEndOf="@id/imageView"
+                app:layout_constraintStart_toStartOf="@id/textViewTitle"
                 app:layout_constraintTop_toBottomOf="@id/textViewTitle"/>
