+++ app/src/main/res/layout/content_scrolling.xml
@@ -25,7 +25,7 @@
             app:layout_collapseMode="parallax"
             app:layout_collapseParallaxMultiplier="0.7">
 
-            <ImageView
+            <androidx.appcompat.widget.AppCompatImageView
                 android:id="@+id/ivContactItem1"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -182,7 +182,7 @@
                 android:textAppearance="@style/TextAppearance.AppCompat.Small"
                 android:textColor="#727272" />
 
-            <ToggleButton
+            <androidx.appcompat.widget.AppCompatToggleButton
                 android:id="@+id/chkState3"
                 android:layout_width="24dp"
                 android:layout_height="24dp"
