+++ app/src/main/res/layout/licence_layout.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
-        android:background="@color/blue">
+        android:background="@color/black">
     <TextView
         android:id="@+id/licenceID"
         android:layout_width="wrap_content"
@@ -38,7 +38,7 @@
         android:layout_height="wrap_content"
         android:textStyle="normal"
         android:text="@string/long_text_example"
-        android:background="#8B42B883"
+        android:background="#8BE5FAF1"
         android:textColor="#000000"
         android:maxLines="4"
         android:textAppearance="?android:textAppearanceSmall"
