+++ app/src/main/res/menu/drawer_view.xml
@@ -34,6 +34,7 @@
                 android:title="@string/nav_about_us" />
             <item
                 android:id="@+id/nav_licences"
+                android:visible="false"
                 android:title="@string/string_licences" />
             <item
                 android:id="@+id/nav_privacy_policy"
